Moved shared export functionality to a concern so it can be used by multiple export-related controllers
